refactored useRandomIndex into useRandomPokemon hook
Main change here is to just call Math.random() during the event handler
and use that fixed random value in the setState updater (this is
deterministic).

This also avoids needing a shuffle function. This diff just maintains
a separate array for questions that should be asked, have been asked,
and were answered correctly in a streak.